Thank you - database changes, plus a small css fix.
To get the thankyou system working in the database, you need the links table and the thankyou template inserted into the terms table. The links table stores the hashes used in the Accept/Decline email sent to the end user. Clicking on those urls performs an action, thus the hashes are needed to prevent malicious thanking.